ht_malloc.c: fallback to mmap for recursion
There is a problem: when dlsym wants to call malloc (or calloc),
it calls hyperthymesia wrappers, which sometimes also want call dlsym,
especially in cases of finding 'real' libc functions.

This leads to infinite recursion, and breaks the app.

This patch makes an fallback of malloc/realloc to mmap allocator for
not-managed (e.g. recursive) calls.

mmap is much slower, but works OK in rare cases of dlsym in malloc.

Signed-off-by: rkhapov <r.khapov@ya.ru>